backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
psycopg2-binary
python-dotenvs